+++ src/pycookiecheat/__main__.py
@@ -16,7 +16,7 @@ def main() -> None:
         prog="pycookiecheat",
         description="Copy cookies from Chrome or Firefox and output as json",
     )
-    parser.add_argument("url", help="requires scheme (e.g. `https://`)")
+    parser.add_argument("url")
     parser.add_argument("-b", "--browser", default="Chrome")
     parser.add_argument(
         "-o",